+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add the ability to un-ban users from circles, groups, and galleries
 * Reduce size of the circle avatar, so "Community" fits on one line on iPad
 * Add ability to rename circles, groups, and galleries
+* Add ability to change group topic
 
 # v1.0.3
 * Bugfix: Ignore invalid config object in account data